HSX parse without fixities
This is a much more satisfying solution than having a hardcoded list of fixities.
One drawback is that the parse result will be wrong, so pretty printing there is
a chance that code might break / have a different answer. However in some
preliminary testing that doesn't seem to be the case.